Feat: 열쇠 및 금고 추가
금고 안에 무기를 넣고 열쇠로 여는 기능은 추후 구현 예정

+++ CODE_PANDEMIC/Assets/Scripts/Puzzle/Interact/Fire Extinguisher/PZ_FireExtinguisher_Effect.cs
@@ -6,6 +6,6 @@ public class PZ_FireExtinguisher_Effect : MonoBehaviour
 
     private void OnTriggerEnter(Collider other)
     {
-        other.GetComponent<AI_Controller>()?.TakeDamage(_damage);
+        other.GetComponent<AI_Controller>()?.TakeDamage((int)_damage);
     }
 }
